Melhorias e correções exercícios aula 19 + Reorganização exercícios aula 13

+++ src/exercicios_aula_19/Exer4.java
@@ -11,7 +11,7 @@ public class Exer4 {
         //    4. Criar um vetor A com 15 elementos inteiros. Construir um vetor B de mesmo tamanho, sendo que cada elemento do vetor B 
         //       deverá ser a raiz quadrada do respectivo elemento de A, ou seja: B[i] = sqrt(A[i]).
         
-        double[] A = {11, 2, 9, 65, 14, 15, 12, 17, 19, 87, 22, 55, 47, 20, 10};
+        int[] A = {11, 2, 9, 65, 14, 15, 12, 17, 19, 87, 22, 55, 47, 20, 10};
         double[] B = new double[A.length];
 
         for (int i = 0; i < A.length; i++) {
